:shirt: All compiler warnings are fixed.
