+++ array.h
@@ -26,4 +26,5 @@ bool apush();
 long alen();
 ARRAY *anew();
 void aunshift(register ARRAY *, register int);
+void ajoin(register ARRAY *, char *, register STR *);
