+++ str.h
@@ -39,4 +39,8 @@ void str_sset(STR *, register STR *);
 void str_set(register STR *, register char *);
 void str_numset(register STR *, double);
 void str_chop(register STR *, register char *);
+void str_inc(register STR *);
+void str_dec(register STR *);
+void str_reset(register char *);
+void str_grow(register STR *, int);
